Refactor form validation into separate util to be reused later.
